chore: get started with viz provider refactor
